Modify recorder and cim_operations for logging
1. Add LogOperationRecorder class to the recorder.  This class logs cim
operations requests/responses and http requests/responses to named
loggers defined by the PywbemLogger class.  It has separate named
loggers for cim_operations and http each of which is defined in
PywbemLogger.
and a set of constants that define the named logger corresponding to
each component (pywbem.ops for the ops component and pywbem.http for the
http component.

2. For some of these logs, an additional detail level entry is defined
that allows setting limits on the log output.  If this log component
attribute is set to 'min' only a defined maximum size is logged for
request responses and http responses.  Not that this is not applied to
requests and http requests because they are really limited in any case.
The only request and http requests that are large are the create and
modify(class/instance) logs and that is normally limited because each
request is just a single CIM Object.

3. Modified cimoperation.py to call these loggers for the request
methods and responses and for http requests and responses. Note that
this code already existed for the TestClientRecorder so there were only
minor modifications.

4. Modified cim_operations.py so that multiple loggers can operate
simultaneously. It now works off a list of active recorders and calls
all of the active recorders.

5. Added a connection id attribute to WBEMConnection that is incremented
for each new connection. This parameter is also included in each log so
that the log can be linked back to the connection that created it. This
id is the combination of an integer that is increment as a class
attribute for each new connection and a process_id to separate logs in
separate processes.

6. Deprecated the original operation_logger setter and added a new
method that adds new loggers to the active logger list.  Made the
original operation_recorder attribute into a property with a getter and
setter. Attached a deprecation warning to both.  The setter sets an
internal variable and also puts the operation logger into the active
logger list.

7. Added a new constructor attribute to WBEMConnection (enable_log) so
that the log can be enabled without having to do it after the connection
is created.

8. Added a property to enable/disable recording so that after recorders
are started they can be interactively enabled and disabled.